requirements.txt
# Core data processing and analysis
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0

# Crypto exchange API (if used for live data)
ccxt>=2.0.0

# Testing
pytest>=6.2.5

# Configuration management
PyYAML>=6.0

# Command-line interface
argparse>=1.4.0
